+++ test/TestAsset/Service.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace ZendTest\AuraDi\Config\TestAsset;
+namespace LaminasTest\AuraDi\Config\TestAsset;
 
 class Service
 {