+++ CHANGELOG.md
@@ -20,7 +20,7 @@
 
 ### Fixed
 
-- [#1](https://github.com/webimpress/zend-auradi-config/pull/1) fixes an issue with lowest
+- [zendframework/zend-auradi-config#1](https://github.com/webimpress/laminas-auradi-config/pull/1) fixes an issue with lowest
   dependencies, when deprecated interop-container was used, instead of PSR-11 container.
 
 ## 0.1.0 - 2017-09-27